Moved output to main module
